result double

+++ CalculatorApp/Calculator/Controllers/HomeController.cs
@@ -11,7 +11,7 @@ namespace Calculator.Controllers
 
         public HomeController()
         {
-            ///_adding = adding;
+            
         }
 
         public IActionResult Index()
@@ -33,22 +33,22 @@ namespace Calculator.Controllers
         [HttpPost]
         public ActionResult Index(string firstNumber, string secondNumber, string Cal)
         {
-            int a = Convert.ToInt32(firstNumber);
-            int b = Convert.ToInt32(secondNumber);
-            float c = 0;
+            double a = Convert.ToDouble(firstNumber);
+            double b = Convert.ToDouble(secondNumber);
+            double c = 0;
             switch (Cal)
             {
                 case "Add":
                     c = _adding.Add(a, b);
                     break;
                 case "Sub":
-                    c = (float)a - b;
+                    c = (double)a - b;
                     break;
                 case "Mul":
-                    c = (float)a * b;
+                    c = (double)a * b;
                     break;
                 case "Div":
-                    c = (float)a / b;
+                    c = (double)a / b;
                     break;
             }
             ViewBag.Result = c;
